Conflicts:
	client/control/input_translator.cpp
	client/control/input_translator.hpp
	client/states/running.cpp

+++ random.cpp
@@ -4,7 +4,12 @@
 sanguis::rand_gen_type
 sanguis::create_seeded_randgen()
 {
-	return rand_gen_type(
-		sge::time::now().time_since_epoch().count()
-	);
+	return
+		rand_gen_type(
+			static_cast<
+				unsigned long
+			>(
+				sge::time::now().time_since_epoch().count()
+			)
+		);
 }
